+++ lib/magma/src/app_driver/magma.cc
@@ -13,22 +13,22 @@
 #include "magma_util/dlog.h"
 #include "magma_util/macros.h"
 
-drm_intel_bo* magma_bo_alloc(drm_intel_bufmgr* bufmgr, const char* name, unsigned long size,
-                             unsigned int alignment)
+drm_intel_bo* magma_bo_alloc(drm_intel_bufmgr* bufmgr, const char* name, uint32_t size,
+                             uint32_t alignment)
 {
     DLOG("magma_bo_alloc %s size %ld alignment 0x%x", name, size, alignment);
     return bufmgr->AllocBufferObject(name, size, alignment, MAGMA_TILING_MODE_NONE, 0 /*stride*/);
 }
 
 drm_intel_bo* magma_bo_alloc_for_render(drm_intel_bufmgr* bufmgr, const char* name,
-                                        unsigned long size, unsigned int alignment)
+                                        uint32_t size, uint32_t alignment)
 {
     UNIMPLEMENTED("magma_bo_alloc_for_render");
     return 0;
 }
 
-drm_intel_bo* magma_bo_alloc_tiled(drm_intel_bufmgr* bufmgr, const char* name, unsigned long size,
-                                   unsigned long flags, uint32_t tiling_mode, unsigned long stride)
+drm_intel_bo* magma_bo_alloc_tiled(drm_intel_bufmgr* bufmgr, const char* name, uint32_t size,
+                                   uint32_t flags, uint32_t tiling_mode, uint32_t stride)
 {
     DLOG("magma_bo_alloc_tiled %s size %lu flags 0x%lx tiling_mode 0x%x stride %lu", name, size,
          flags, tiling_mode, stride);
@@ -36,13 +36,13 @@ drm_intel_bo* magma_bo_alloc_tiled(drm_intel_bufmgr* bufmgr, const char* name, u
     return bufmgr->AllocBufferObject(name, size, 0 /*alignment*/, tiling_mode, stride);
 }
 
-int magma_bo_busy(drm_intel_bo* bo)
+int32_t magma_bo_busy(drm_intel_bo* bo)
 {
     UNIMPLEMENTED("magma_bo_busy");
     return 0;
 }
 
-int magma_bo_emit_reloc(drm_intel_bo* bo, uint32_t offset, drm_intel_bo* target_bo,
+int32_t magma_bo_emit_reloc(drm_intel_bo* bo, uint32_t offset, drm_intel_bo* target_bo,
                         uint32_t target_offset, uint32_t read_domains, uint32_t write_domain)
 {
     DLOG("TODO magma_bo_emit_reloc - offset 0x%x target_offset 0x%x domains 0x%x 0x%x", offset,
@@ -50,38 +50,38 @@ int magma_bo_emit_reloc(drm_intel_bo* bo, uint32_t offset, drm_intel_bo* target_
     return 0;
 }
 
-int magma_bo_flink(drm_intel_bo* bo, uint32_t* name)
+int32_t magma_bo_flink(drm_intel_bo* bo, uint32_t* name)
 {
     UNIMPLEMENTED("mamga_bo_flink");
     return 0;
 }
 
 drm_intel_bo* magma_bo_gem_create_from_name(drm_intel_bufmgr* bufmgr, const char* name,
-                                            unsigned int handle)
+                                            uint32_t handle)
 {
     UNIMPLEMENTED("magma_bo_gem_create_from_name");
     return 0;
 }
 
-drm_intel_bo* magma_bo_gem_create_from_prime(drm_intel_bufmgr* bufmgr, int prime_fd, int size)
+drm_intel_bo* magma_bo_gem_create_from_prime(drm_intel_bufmgr* bufmgr, int32_t prime_fd, int32_t size)
 {
     UNIMPLEMENTED("magma_bo_gem_create_from_prime");
     return 0;
 }
 
-int magma_bo_gem_export_to_prime(drm_intel_bo* bo, int* prime_fd)
+int32_t magma_bo_gem_export_to_prime(drm_intel_bo* bo, int32_t* prime_fd)
 {
     UNIMPLEMENTED("magma_bo_gem_export_to_prime");
     return 0;
 }
 
-int magma_bo_get_subdata(drm_intel_bo* bo, unsigned long offset, unsigned long size, void* data)
+int32_t magma_bo_get_subdata(drm_intel_bo* bo, uint32_t offset, uint32_t size, void* data)
 {
     DLOG("magma_bo_get_subdata '%s' STUB\n", MagmaBuffer::cast(bo)->Name());
     return 0;
 }
 
-int magma_bo_get_tiling(drm_intel_bo* bo, uint32_t* tiling_mode, uint32_t* swizzle_mode)
+int32_t magma_bo_get_tiling(drm_intel_bo* bo, uint32_t* tiling_mode, uint32_t* swizzle_mode)
 {
     DLOG("magma_bo_get_tiling - swizzle stubbed");
     auto buffer = MagmaBuffer::cast(bo);
@@ -91,13 +91,13 @@ int magma_bo_get_tiling(drm_intel_bo* bo, uint32_t* tiling_mode, uint32_t* swizz
     return 0;
 }
 
-int magma_bo_madvise(drm_intel_bo* bo, int madv)
+int32_t magma_bo_madvise(drm_intel_bo* bo, int32_t madv)
 {
     UNIMPLEMENTED("magma_bo_madvise");
     return 0;
 }
 
-int magma_bo_map(drm_intel_bo* bo, int write_enable)
+int32_t magma_bo_map(drm_intel_bo* bo, int32_t write_enable)
 {
     auto buffer = MagmaBuffer::cast(bo);
     DLOG("magma_bo_map %s", buffer->Name());
@@ -105,8 +105,8 @@ int magma_bo_map(drm_intel_bo* bo, int write_enable)
     return 0;
 }
 
-int magma_bo_mrb_exec(drm_intel_bo* bo, int used, void* unused, int num_cliprects, int DR4,
-                      unsigned int flags)
+int32_t magma_bo_mrb_exec(drm_intel_bo* bo, int32_t used, void* unused, int32_t num_cliprects, int32_t DR4,
+                      uint32_t flags)
 {
     UNIMPLEMENTED("magma_bo_mrb_exec");
     return 0;
@@ -119,7 +119,7 @@ void magma_bo_reference(drm_intel_bo* bo)
     buffer->Incref();
 }
 
-int magma_bo_references(drm_intel_bo* bo, drm_intel_bo* target_bo)
+int32_t magma_bo_references(drm_intel_bo* bo, drm_intel_bo* target_bo)
 {
     auto buffer = MagmaBuffer::cast(bo);
     auto target_buffer = MagmaBuffer::cast(target_bo);
@@ -128,13 +128,13 @@ int magma_bo_references(drm_intel_bo* bo, drm_intel_bo* target_bo)
     return 0;
 }
 
-int magma_bo_subdata(drm_intel_bo* bo, unsigned long offset, unsigned long size, const void* data)
+int32_t magma_bo_subdata(drm_intel_bo* bo, uint32_t offset, uint32_t size, const void* data)
 {
     DLOG("magma_bo_subdata '%s' STUB", MagmaBuffer::cast(bo)->Name());
     return 0;
 }
 
-int magma_bo_unmap(drm_intel_bo* bo)
+int32_t magma_bo_unmap(drm_intel_bo* bo)
 {
     auto buffer = MagmaBuffer::cast(bo);
     DLOG("magma_bo_unmap %p", bo);
@@ -158,7 +158,7 @@ void magma_bo_wait_rendering(drm_intel_bo* bo)
     buffer->WaitRendering();
 }
 
-int magma_bufmgr_check_aperture_space(drm_intel_bo** bo_array, int count)
+int32_t magma_bufmgr_check_aperture_space(drm_intel_bo** bo_array, int32_t count)
 {
     DLOG("magma_bufmgr_check_aperture_space - STUB");
     return 0;
@@ -180,15 +180,15 @@ void magma_bufmgr_gem_enable_reuse(drm_intel_bufmgr* bufmgr)
     DLOG("magma_bufmgr_gem_enable_reuse - STUB");
 }
 
-int magma_bufmgr_gem_get_devid(drm_intel_bufmgr* bufmgr)
+int32_t magma_bufmgr_gem_get_devid(drm_intel_bufmgr* bufmgr)
 {
     DLOG("magma_bufmgr_gem_get_devid");
-    int id = static_cast<int>(bufmgr->GetDeviceId());
+    int32_t id = static_cast<int>(bufmgr->GetDeviceId());
     DLOG("returning id 0x%x", id);
     return id;
 }
 
-drm_intel_bufmgr* magma_bufmgr_gem_init(int device_handle, int batch_size)
+drm_intel_bufmgr* magma_bufmgr_gem_init(int32_t device_handle, int32_t batch_size)
 {
     DLOG("magma_bufmgr_gem_init device_handle 0x%x batch_size %d", device_handle, batch_size);
 
@@ -206,12 +206,12 @@ void magma_bufmgr_gem_set_aub_annotations(drm_intel_bo* bo, drm_intel_aub_annota
     UNIMPLEMENTED("magma_bufmgr_gem_set_aub_annotations");
 }
 
-void magma_bufmgr_gem_set_aub_dump(drm_intel_bufmgr* bufmgr, int enable)
+void magma_bufmgr_gem_set_aub_dump(drm_intel_bufmgr* bufmgr, int32_t enable)
 {
     UNIMPLEMENTED("magma_bufmgr_gem_set_aub_dump");
 }
 
-void magma_bufmgr_set_debug(drm_intel_bufmgr* bufmgr, int enable_debug)
+void magma_bufmgr_set_debug(drm_intel_bufmgr* bufmgr, int32_t enable_debug)
 {
     UNIMPLEMENTED("magma_bufmgr_set_debug");
 }
@@ -230,7 +230,7 @@ void magma_decode_context_free(struct drm_intel_decode* ctx)
 }
 
 void magma_decode_set_batch_pointer(struct drm_intel_decode* ctx, void* data, uint32_t hw_offset,
-                                    int count)
+                                    int32_t count)
 {
     UNIMPLEMENTED("magma_decode_set_batch_pointer");
 }
@@ -240,37 +240,37 @@ void magma_decode_set_output_file(struct drm_intel_decode* ctx, FILE* out)
     UNIMPLEMENTED("magma_decode_set_output_file");
 }
 
-void magma_gem_bo_aub_dump_bmp(drm_intel_bo* bo, int x1, int y1, int width, int height,
-                               enum aub_dump_bmp_format format, int pitch, int offset)
+void magma_gem_bo_aub_dump_bmp(drm_intel_bo* bo, int32_t x1, int32_t y1, int32_t width, int32_t height,
+                               enum aub_dump_bmp_format format, int32_t pitch, int32_t offset)
 {
     UNIMPLEMENTED("magma_gem_bo_aub_dump_bmp");
 }
 
-void magma_gem_bo_clear_relocs(drm_intel_bo* bo, int start)
+void magma_gem_bo_clear_relocs(drm_intel_bo* bo, int32_t start)
 {
     DLOG("magma_gem_bo_clear_relocs - STUB");
 }
 
-int magma_gem_bo_get_reloc_count(drm_intel_bo* bo)
+int32_t magma_gem_bo_get_reloc_count(drm_intel_bo* bo)
 {
     DLOG("magma_gem_bo_get_reloc_count - STUB");
     return 0;
 }
 
-int magma_gem_bo_map_gtt(drm_intel_bo* bo)
+int32_t magma_gem_bo_map_gtt(drm_intel_bo* bo)
 {
     UNIMPLEMENTED("magma_gem_bo_map_gtt");
     return 0;
 }
 
-int magma_gem_bo_map_unsynchronized(drm_intel_bo* bo)
+int32_t magma_gem_bo_map_unsynchronized(drm_intel_bo* bo)
 {
     DLOG("magma_gem_bo_map_unsynchronized %s - using regular map", MagmaBuffer::cast(bo)->Name());
-    const int write_enable = 1;
+    const int32_t write_enable = 1;
     return magma_bo_map(bo, write_enable);
 }
 
-int magma_gem_bo_wait(drm_intel_bo* bo, int64_t timeout_ns)
+int32_t magma_gem_bo_wait(drm_intel_bo* bo, int64_t timeout_ns)
 {
     UNIMPLEMENTED("magma_gem_bo_wait");
     return 0;
@@ -289,31 +289,31 @@ drm_intel_context* magma_gem_context_create(drm_intel_bufmgr* bufmgr)
 
 void magma_gem_context_destroy(drm_intel_context* ctx) { DLOG("magma_gem_context_destroy - STUB"); }
 
-int magma_gem_bo_context_exec(drm_intel_bo* bo, drm_intel_context* ctx, int used,
-                              unsigned int flags)
+int32_t magma_gem_bo_context_exec(drm_intel_bo* bo, drm_intel_context* ctx, int32_t used,
+                              uint32_t flags)
 {
     auto buffer = MagmaBuffer::cast(bo);
-    int context_id = static_cast<int>(reinterpret_cast<intptr_t>(ctx));
+    int32_t context_id = static_cast<int>(reinterpret_cast<intptr_t>(ctx));
     DLOG("magma_gem_bo_context_exec buffer '%s' context_id %d", buffer->Name(), context_id);
-    // int int_context_id = static_cast<int>(context_id);
+    // int32_t int_context_id = static_cast<int>(context_id);
     buffer->device()->ExecuteBuffer(buffer, context_id, used, flags);
     return 0;
 }
 
-int magma_get_aperture_sizes(int fd, size_t* mappable, size_t* total)
+int32_t magma_get_aperture_sizes(int32_t fd, size_t* mappable, size_t* total)
 {
     UNIMPLEMENTED("magma_get_aperture_sizes");
     return 0;
 }
 
-int magma_get_reset_stats(drm_intel_context* ctx, uint32_t* reset_count, uint32_t* active,
+int32_t magma_get_reset_stats(drm_intel_context* ctx, uint32_t* reset_count, uint32_t* active,
                           uint32_t* pending)
 {
     UNIMPLEMENTED("magma_get_reset_stats");
     return 0;
 }
 
-int magma_reg_read(drm_intel_bufmgr* bufmgr, uint32_t offset, uint64_t* result)
+int32_t magma_reg_read(drm_intel_bufmgr* bufmgr, uint32_t offset, uint64_t* result)
 {
     DLOG("magma_reg_read - STUB returning 0");
     return 0;
